pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hs2p"
version = "3.3.0"
description = "Optimized slide tiling library for histopathology"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
    { name = "Clément Grisi", email = "clement.grisi@radboudumc.nl" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "omegaconf",
    "numpy",
    "pandas",
    "h5py",
    "tqdm",
    "rich",
    "wandb",
    "pillow",
    "matplotlib",
    "opencv-python",
]

[project.urls]
Homepage = "https://github.com/clemsgrs/hs2p"
"Bug Tracker" = "https://github.com/clemsgrs/hs2p/issues"

[project.scripts]
hs2p = "hs2p.__main__:entrypoint"

[project.optional-dependencies]
openslide = [
    "openslide-python",
    "openslide-bin",
]
asap = [
    "wholeslidedata==0.0.15",
    "numpy<2",
]
cucim = [
    "cucim-cu12>=26.02.00",
    "cupy-cuda12x",
    "nvidia-nvimgcodec-cu12>=0.7.0",
]
vips = [
    "pyvips",
]
all = [
    "openslide-python",
    "openslide-bin",
    "wholeslidedata==0.0.15",
    "numpy<2",
    "cucim-cu12>=26.02.00",
    "cupy-cuda12x",
    "nvidia-nvimgcodec-cu12>=0.7.0",
    "pyvips",
]
testing = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "mypy>=0.910",
    "flake8>=3.9",
    "flake8-pyproject>=1.2.3",
    "tox>=3.24",
]
sam2 = [
    "torch>=2.0.0",
    "hydra-core>=1.3.2",
    "huggingface-hub>=0.23.0",
]

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.packages.find]
include = ["hs2p*"]

[tool.setuptools.package-data]
hs2p = ["py.typed", "configs/*.yaml"]

[tool.pytest.ini_options]
addopts = "--cov=hs2p -m 'not integration and not script'"
testpaths = [
    "tests",
]
markers = [
    "integration: backend-dependent or fixture-dependent integration coverage excluded from the default fast test run",
    "script: tests for standalone or local-only scripts excluded from the default CI run",
]

[tool.mypy]
mypy_path = "."
check_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
no_implicit_reexport = true

[tool.flake8]
max-line-length = 160

[tool.bumpver]
current_version = "3.3.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit = false       # We do version bumping in CI, not as a commit
tag = false          # Git tag already exists — we don't auto-tag
push = false         # Don't push anything in CI
files = [
    "pyproject.toml",
    "hs2p/__init__.py"
]

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    '^current_version = "{version}"$',
    '^version = "{version}"$',
]
"hs2p/__init__.py" = [
    '^__version__ = "{version}"$',
]
